refact: Decouple model downloading from generation module
